improve sound

+++ R-Type/client/includes/GUI/IGraphic.hpp
@@ -10,7 +10,7 @@ class IGraphic
 
 	// public methods
 	public:
-		virtual bool		drawSprite(std::string const &key, float delta, float x, float y) = 0;
+		virtual bool		drawSprite(std::string const &key, uint32_t delta, float x, float y) = 0;
 		virtual bool		drawFont(std::string const &key, std::string const &strn, float x, float y, uint32_t size = 12) = 0;
 		virtual bool		playSound(std::string const &key, bool onLoop = false) = 0;
 		virtual bool		playMusic(std::string const &key, bool onLoop = true) = 0;
